Card Flipping working better
The issue is that the buttons were registering when they weren't supposed to, both during time.sleep and while in the flipping animation. However these clicks could not be executed on since the process was busy, thus it saved the clicks and then loaded them all at once. This both added a whole bunch of  cards to the selection table (why the card would remain shown) and to the flipping process (why it would flip a whole bunch)